_archive/deprecated/test_multi_account_quick.py
#!/usr/bin/env python3
"""
Quick test for multi-account system without full dependency chain.
"""

import asyncio
import tempfile
import os
import yaml
from datetime import datetime

# Test the core account manager without dependencies
class MockAccountManager:
    """Mock account manager for testing"""
    
    def __init__(self, config_path):
        self.config_path = config_path
        self.accounts = {}
        self.current_account = None
    
    async def initialize(self):
        """Initialize the account manager"""
        await self._load_registry()
        await self._set_default_account()
        print(f"✓ Mock account manager initialized with {len(self.accounts)} accounts")
    
    async def _load_registry(self):
        """Load account registry from YAML file"""
        if not os.path.exists(self.config_path):
            print(f"⚠ Config file not found: {self.config_path}")
            return
        
        try:
            with open(self.config_path, 'r') as f:
                registry_data = yaml.safe_load(f)
            
            if not registry_data:
                return
            
            # Load accounts
            for account_data in registry_data.get("accounts", []):
                account_id = account_data["id"]
                self.accounts[account_id] = account_data
            
            print(f"✓ Loaded {len(self.accounts)} accounts from registry")
            
        except Exception as e:
            print(f"✗ Failed to load account registry: {e}")
    
    async def _set_default_account(self):
        """Set default account"""
        if not self.accounts:
            return
        
        # Get active accounts
        active_accounts = [
            acc for acc in self.accounts.values()
            if acc.get("status") == "active"
        ]
        
        if active_accounts:
            self.current_account = active_accounts[0]["id"]
            print(f"✓ Set default account: {self.current_account}")
    
    def get_current_account(self):
        """Get current account"""
        return self.accounts.get(self.current_account) if self.current_account else None
    
    async def switch_account(self, account_id):
        """Switch account"""
        if account_id not in self.accounts:
            print(f"✗ Account not found: {account_id}")
            return False
        
        account = self.accounts[account_id]
        if account.get("status") != "active":
            print(f"✗ Account {account_id} is not active: {account.get('status')}")
            return False
        
        old_account = self.current_account
        self.current_account = account_id
        print(f"✓ Switched from {old_account} to {account_id}")
        return True
    
    async def create_account(self, name, account_type, email, provider, quota_limit, models_preferred, priority):
        """Create new account"""
        account_id = f"{provider}_{name.replace(' ', '_').lower()}_{len(self.accounts) + 1}"
        
        account_data = {
            "id": account_id,
            "name": name,
            "type": account_type,
            "status": "active",
            "created_at": datetime.now().isoformat(),
            "last_used": None,
            "email": email,
            "provider": provider,
            "quota_remaining": quota_limit,
            "quota_limit": quota_limit,
            "models_preferred": models_preferred,
            "priority": priority,
            "api_key": None,
            "usage_stats": {
                "total_requests": 0,
                "successful_requests": 0,
                "failed_requests": 0,
                "avg_response_time": 0.0
            }
        }
        
        self.accounts[account_id] = account_data
        print(f"✓ Created account: {account_id}")
        return account_id


async def test_multi_account_system():
    """Test the multi-account system"""
    print("🚀 Testing Multi-Account System")
    print("=" * 50)
    
    # Create test configuration
    with tempfile.TemporaryDirectory() as temp_dir:
        config_path = os.path.join(temp_dir, "test_accounts.yaml")
        
        test_accounts = {
            "accounts": [
                {
                    "id": "github_user_01",
                    "name": "GitHub User Account",
                    "type": "user",
                    "status": "active",
                    "created_at": datetime.now().isoformat(),
                    "last_used": None,
                    "email": "user@example.com",
                    "provider": "github",
                    "quota_remaining": 1000,
                    "quota_limit": 1000,
                    "models_preferred": ["haiku-4.5", "sonnet-4.5"],
                    "priority": 1,
                    "api_key": "test_key_123",
                    "usage_stats": {
                        "total_requests": 0,
                        "successful_requests": 0,
                        "failed_requests": 0,
                        "avg_response_time": 0.0
                    }
                },
                {
                    "id": "gemini_user_01",
                    "name": "Gemini User Account",
                    "type": "user",
                    "status": "active",
                    "created_at": datetime.now().isoformat(),
                    "last_used": None,
                    "email": "user@example.com",
                    "provider": "gemini",
                    "quota_remaining": 800,
                    "quota_limit": 1000,
                    "models_preferred": ["gemini-3-pro"],
                    "priority": 2,
                    "api_key": "test_key_456",
                    "usage_stats": {
                        "total_requests": 0,
                        "successful_requests": 0,
                        "failed_requests": 0,
                        "avg_response_time": 0.0
                    }
                }
            ],
            "rotation_strategy": "quota_aware",
            "auto_switch": True,
            "validation_enabled": True
        }
        
        # Write test config
        with open(config_path, 'w') as f:
            yaml.dump(test_accounts, f)
        
        print(f"📁 Created test configuration: {config_path}")
        
        # Test account manager
        print("\n🔧 Testing Account Manager")
        print("-" * 30)
        
        manager = MockAccountManager(config_path)
        await manager.initialize()
        
        # Test getting current account
        current = manager.get_current_account()
        print(f"✓ Current account: {current['name'] if current else 'None'}")
        
        # Test switching accounts
        success = await manager.switch_account("gemini_user_01")
        print(f"✓ Account switch: {'Success' if success else 'Failed'}")
        
        # Test creating new account
        new_account_id = await manager.create_account(
            name="New Test Account",
            account_type="user",
            email="new@example.com",
            provider="opencode",
            quota_limit=2000,
            models_preferred=["raptor-mini"],
            priority=1
        )
        
        # Test switching to new account
        success = await manager.switch_account(new_account_id)
        print(f"✓ Switch to new account: {'Success' if success else 'Failed'}")
        
        print("\n✅ All tests passed!")
        print("\n📊 Test Summary:")
        print(f"   - Accounts loaded: {len(manager.accounts)}")
        print(f"   - Current account: {manager.current_account}")
        print(f"   - Account switching: Working")
        print(f"   - Account creation: Working")
        
        return True


def main():
    """Main test function"""
    try:
        success = asyncio.run(test_multi_account_system())
        if success:
            print("\n🎉 Multi-Account System Test: PASSED")
            return 0
        else:
            print("\n❌ Multi-Account System Test: FAILED")
            return 1
    except Exception as e:
        print(f"\n💥 Test failed with error: {e}")
        return 1


if __name__ == "__main__":
    exit(main())